Fix traversal weight task definition and queue clearing
